+++ Core/Graphics/SpriteTexture.cs
@@ -100,7 +100,7 @@ public class SpriteTexture
         spriteBatch.Draw(Texture, position, Clip, color, 0, -Origin, 1f, SpriteEffects.None, 0);
     }
 
-    public void DrawTexture(SpriteBatch spriteBatch, Vector2 position, Color color, float rotation, Vector2 scale, SpriteEffects spriteEffects, int zIndex) 
+    public void DrawTexture(SpriteBatch spriteBatch, Vector2 position, Color color, float rotation, Vector2 scale, SpriteEffects spriteEffects, float zIndex) 
     {
         var col = color;
 #if DEBUG
@@ -110,7 +110,7 @@ public class SpriteTexture
         spriteBatch.Draw(Texture, position, Clip, col, rotation, -Origin, scale, spriteEffects, zIndex);
     }
 
-    public void DrawTexture(SpriteBatch spriteBatch, Vector2 position, Color color, float rotation, Vector2 offset, Vector2 scale, SpriteEffects spriteEffects, int zIndex) 
+    public void DrawTexture(SpriteBatch spriteBatch, Vector2 position, Color color, float rotation, Vector2 offset, Vector2 scale, SpriteEffects spriteEffects, float zIndex) 
     {
         var col = color;
 #if DEBUG
@@ -120,7 +120,7 @@ public class SpriteTexture
         spriteBatch.Draw(Texture, position, Clip, col, rotation, offset, scale, spriteEffects, zIndex);
     }
 
-    public void DrawTexture(SpriteBatch spriteBatch, Vector2 position, Rectangle rectangle, Color color, float rotation, Vector2 offset, Vector2 scale, SpriteEffects spriteEffects, int zIndex) 
+    public void DrawTexture(SpriteBatch spriteBatch, Vector2 position, Rectangle rectangle, Color color, float rotation, Vector2 offset, Vector2 scale, SpriteEffects spriteEffects, float zIndex) 
     {
         var col = color;
 #if DEBUG
@@ -130,7 +130,7 @@ public class SpriteTexture
         spriteBatch.Draw(Texture, position, rectangle, col, rotation, offset, scale, spriteEffects, zIndex);
     }
 
-    public void DrawTexture(SpriteBatch spriteBatch, Vector2 position, Rectangle rectangle, Color color, float rotation, Vector2 offset, float scale, SpriteEffects spriteEffects, int zIndex) 
+    public void DrawTexture(SpriteBatch spriteBatch, Vector2 position, Rectangle rectangle, Color color, float rotation, Vector2 offset, float scale, SpriteEffects spriteEffects, float zIndex) 
     {
         var col = color;
 #if DEBUG
